requirements.txt
Django>=1.11,<1.12
djangorestframework>=3.3.3
wagtail>=1.10
wagtailmenus
django-bootstrap3>=7.1.0
django-libsass>=0.7
django-modelcluster>=3.0
django-overextends>=0.4.2
django-storage-swift
elasticsearch>=1.2.0
libsass>=0.8.2
Pillow>=3.2.0
wagtail-django-recaptcha>=0.1
psycopg2>=2.6
mcstatus>=2.1
psutil>=4.1.0
markdown>=2.6.8
pymdown-extensions>=3.4
